Java Project based on swing.
This application  allows ,for user profiles to be added to, deleted
from, or             looked-up in the social network. Moreover, for each
profile, It  keeps  track of the person's name associated with that
profile, an optional image that the person may wish to display with
his/her profile, an optional "current status" for the profile currently
engaged in, and a list of friends for each profile.